Fix: Remove border from ship sprite
When creating the sprite from the spritesheet it seems like you need to specify the pixel that is the first pixel within the sprite (within the border) and subtract the size of it. Not sure why the size should be adjusted since within the border it is 16 x16, but only 14x14 will remove the border from all sides.