Fix: Aliens x movement is now handled as grid
Instead of every single alien handling their own movement in x direction update, this is now done in a single dimensional grid that keeps track of the position and updates it for each column. Then the alien only needs to check which position his column is at and take that value.