Add: Alien now moves in y-axis if he isn't in the right position
This adds the functionality that the alien flies downwars until he is in its original position. For testing purposes the boolean isRespawning was added to its descriptor. This should probably be removed when the respawning functionality is created. I think it would be easier to just move the enemy when he is off-screen instead of deleting and creating a new one.